feat: Rewrite monitor to support new strategy
