+++ package.json
@@ -12,8 +12,7 @@
     "migrate": "./migrations/migrate",
     "renew-blockchain-data": "./scripts/renewBlockchainData.js",
     "seed": "./scripts/seed.js",
-    "watch-marketplace": "./scripts/monitor/index.js transform Marketplace AuctionCreated AuctionSuccessful AuctionCancelled",
-    "watch-land": "./scripts/monitor/index.js transform LANDRegistry Update Transfer",
+    "monitor-blockchain": "./scripts/monitor/index.js store",
     "build": "npm run lint && npm run test",
     "start": "cd src && babel-node server.js",
     "dev": "nodemon --exec npm run start"
