+++ scripts/monitor/handlers/index.js
@@ -1,115 +1,34 @@
-// Check the default file search progression on './HandlerIndex#get'
+import { Log } from 'decentraland-commons'
+import { BlockchainEvent } from '../../../src/BlockchainEvent'
 
-import { eth, txUtils, contracts } from 'decentraland-commons'
-import { decodeAssetId, debounceById } from './utils'
-import { Parcel } from '../../../src/Parcel'
-import { Publication } from '../../../src/Publication'
+const log = new Log('handlers')
 
-export async function transform_Marketplace(event) {
-  if (event.removed) return
+export async function store(eventData) {
+  if (eventData.removed) return
 
-  const { transactionHash } = event
-  const { assetId } = event.args
-  const id = await decodeAssetId(assetId)
+  const { event, transactionHash, blockNumber } = eventData
 
-  switch (event.event) {
-    case 'AuctionCreated': {
-      const { creator, priceInWei, expiresAt } = event.args
-      const [x, y] = Parcel.splitId(id)
-
-      console.log(
-        `[Marketplace-AuctionCreated] Creating publication for ${x},${y}`
-      )
-
-      await Publication.insert({
-        tx_hash: transactionHash,
-        tx_status: txUtils.TRANSACTION_STATUS.confirmed,
-        status: Publication.STATUS.open,
-        owner: creator.toLowerCase(),
-        buyer: null,
-        price: eth.utils.fromWei(priceInWei.toNumber()),
-        expires_at: new Date(expiresAt.toNumber()),
-        x,
-        y
-      })
-      break
-    }
-    case 'AuctionSuccessful': {
-      const { totalPrice, winner } = event.args
-
-      console.log(
-        `[Marketplace-AuctionSuccessful] Publication ${id} sold to ${winner}`
-      )
-
-      await Publication.update(
-        {
-          status: Publication.STATUS.sold,
-          buyer: winner.toLowerCase(),
-          price: eth.utils.fromWei(totalPrice.toNumber())
-        },
-        { id }
-      )
-      await Parcel.update({ owner: winner }, { id })
-      break
-    }
-    case 'AuctionCancelled': {
-      console.log(`[Marketplace-AuctionCancelled] Publication ${id} cancelled`)
-
-      await Publication.update({ status: Publication.STATUS.cancelled }, { id })
-      break
-    }
-    default:
-      break
+  const exists = await BlockchainEvent.count({ tx_hash: transactionHash })
+  if (exists) {
+    log.info(`[${event}] Blockchain event ${transactionHash} already exists`)
+    return
   }
+  log.info(`[${event}] Storing blockchain event`)
 
-  return event
-}
+  const args = Object.keys(eventData.args).reduce(
+    (memo, key) =>
+      Object.assign(memo, {
+        [key]: eventData.args[key].toString()
+      }),
+    {}
+  )
 
-export async function transform_LANDRegistry(event) {
-  if (event.removed) return
-
-  const { assetId } = event.args
-  const id = await decodeAssetId(assetId)
-
-  switch (event.event) {
-    case 'Update': {
-      try {
-        const { data } = event.args
-        const attributes = { data: contracts.LANDRegistry.decodeLandData(data) }
-
-        debounceById(id, () => {
-          const attrsStr = JSON.stringify(attributes)
-          console.log(`[LANDRegistry-Update] Updating "${id}" with ${attrsStr}`)
-
-          Parcel.update(attributes, { id })
-        })
-      } catch (error) {
-        // Skip badly formed data
-      }
-      break
-    }
-    case 'Transfer': {
-      const { to } = event.args
-      const [x, y] = Parcel.splitId(id)
-
-      debounceById(id, () => {
-        console.log(
-          `[LANDRegistry-Transfer] Updating "${id}" owner with "${to}"`
-        )
-        Publication.update(
-          { status: Publication.STATUS.cancelled },
-          {
-            x,
-            y
-          }
-        )
-        Parcel.update({ owner: to.toLowerCase() }, { id })
-      })
-      break
-    }
-    default:
-      break
-  }
+  await BlockchainEvent.insert({
+    tx_hash: transactionHash,
+    name: event,
+    block_number: blockNumber,
+    args
+  })
 
   return event
 }
