+++ scripts/monitor/utils.js
